Add API for system color preference reading
Virtually all of our target systems/environments have a way to enable a
system-wide dark mode.

This adds API to read and monitor these settings. This allows for
example to automatically switch between two KColorSchemes (e.g. Breeze
and Breeze Dark). However the API is intentionally independant from
Breeze and KColorScheme.

It contains a backend for Windows based on the existing code in
KColorSchemeManager and a Linux implementation based on XDG portals.