keysequence: Fix warning message about sequence size to be more precise
Sequence of length/count equal to MaxKeyCount is perfectly valid. The
only invalid thing about it is an attempt to actually append more keys
to it.

This message is erroneously being triggered when typing a key sequence
of max length 4 in a KeySequenceItem.qml / KeySequenceHelper: when final
fourth shortcut is pressed a recorder should immediately abort
listening, but instead it tries to append again the modifiers which are
still pressed on a keyboard, resulting in a meaningless warning in
console output.