keysequence: Fix race between recording and currentKeySequence
It used to trigger an assert/warning in appendToSequence because of a
race condition when currentKeySequenceChanged() signal is already
propagating, so that a new value of currentKeySequence() is being
queried from e.g. QML side, but the recording still technically hasn't
finished, so m_isRecording is true, and thus an additional check is
needed to prevent an attempt to construct a key sequence of enormously
large length.

Related to (amends in its own way) commit 92185ac6c40fbbb5dc24832e496539b10c546972